+++ lib/rets/client.rb
@@ -470,7 +470,7 @@ EOF
     end
 
     def binary?(data)
-      data.slice(0, 1024).find {|b| b >= "\x0" && b < " " && b != '-' && b != '~' && b != "\t" && b != "\r" && b != "\n"}
+      data.slice(0, 1024).any? {|b| b >= "\x0" && b < " " && b != '-' && b != '~' && b != "\t" && b != "\r" && b != "\n"}
     end
 
     def tries