+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-### 0.2.2
+### 0.3.1.dev
 
-* WIP
+* correctly handle digest authentication
+
+### 0.3.0.rc.0 / 2012-07-26
+
+* feature: significantly better handling of authorization failures
 
 ### 0.2.1 / 2012-04-20